reorganize sort headers
